% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeBIC}
\alias{computeBIC}
\title{FUNCTION_TITLE}
\usage{
computeBIC(params)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
