% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correctReadCounts}
\alias{correctReadCounts}
\title{FUNCTION_TITLE}
\usage{
correctReadCounts(
  x,
  chrNormalize = c(1:22),
  mappability = 0.9,
  samplesize = 50000,
  verbose = TRUE
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{chrNormalize}{PARAM_DESCRIPTION, Default: c(1:22)}

\item{mappability}{PARAM_DESCRIPTION, Default: 0.9}

\item{samplesize}{PARAM_DESCRIPTION, Default: 50000}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
