% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPanelOfNormals.R
\name{createPanelOfNormals}
\alias{createPanelOfNormals}
\title{Runs createPanelOfNormals for ichorCNA}
\usage{
createPanelOfNormals(
  gcWig,
  mapWig,
  repTimeWig = NULL,
  filelist,
  outfile,
  centromere,
  flankLength = 1e+05,
  chrs = "c(1:22,\\"X\\")",
  genomeStyle = "NCBI",
  genomeBuild = "hg19",
  chrNormalize = "c(1:22)",
  minMapScore = 0,
  maleChrXLogRThres = -0.8,
  fracReadsInChrYForMale = 0.001,
  exons.bed = NULL,
  method = "median",
  ylim = "c(-2,2)",
  plotChrPanels = FALSE
)
}
\arguments{
\item{gcWig}{GC Wig file for reference genome}

\item{mapWig}{Mappabiliy Wig file for reference genome}

\item{repTimeWig}{type = "character", default=NULL, help ="Path to replication timing WIG file.}

\item{filelist}{List of of wig files.}

\item{outfile}{Output file.}

\item{centromere}{File containing Centromere locations}

\item{flankLength}{Length of region flanking centromere to remove.}

\item{chrs}{Specify chromosomes to analyze.}

\item{genomeStyle}{NCBI or UCSC chromosome naming convention; use UCSC if desired output is to have \"chr\" string.}

\item{genomeBuild}{Geome build.}

\item{chrNormalize}{Specify chromosomes to normalize GC/mappability biases}

\item{minMapScore}{Include bins with a minimum mappability score of this value.}

\item{maleChrXLogRThres}{ChrX Log ratio threshold to confirm as male gender.}

\item{fracReadsInChrYForMale}{Threshold for fraction of reads in chrY to assign as male.}

\item{exons.bed}{Path to bed file containing exon regions.}

\item{method}{Median or Mean.}

\item{ylim}{Y-limits for plotting of mean/median log ratios}

\item{plotChrPanels}{Plot PoN values.}
}
\description{
Runs createPanelOfNormals for ichorCNA
}
