% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{dirichletpdf}
\alias{dirichletpdf}
\title{FUNCTION_TITLE}
\usage{
dirichletpdf(x, alpha)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
