% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{estimateMixWeightsParamMap}
\alias{estimateMixWeightsParamMap}
\title{FUNCTION_TITLE}
\usage{
estimateMixWeightsParamMap(rho, kappa)
}
\arguments{
\item{rho}{PARAM_DESCRIPTION}

\item{kappa}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
