% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{estimateParamsMap}
\alias{estimateParamsMap}
\title{FUNCTION_TITLE}
\usage{
estimateParamsMap(
  D,
  n_prev,
  sp_prev,
  phi_prev,
  lambda_prev,
  pi_prev,
  A_prev,
  params,
  rho,
  Zcounts,
  estimateNormal = TRUE,
  estimatePloidy = TRUE,
  estimatePrecision = TRUE,
  estimateInitDist = TRUE,
  estimateTransition = TRUE,
  estimateSubclone = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{D}{PARAM_DESCRIPTION}

\item{n_prev}{PARAM_DESCRIPTION}

\item{sp_prev}{PARAM_DESCRIPTION}

\item{phi_prev}{PARAM_DESCRIPTION}

\item{lambda_prev}{PARAM_DESCRIPTION}

\item{pi_prev}{PARAM_DESCRIPTION}

\item{A_prev}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{rho}{PARAM_DESCRIPTION}

\item{Zcounts}{PARAM_DESCRIPTION}

\item{estimateNormal}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePloidy}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePrecision}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateInitDist}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateTransition}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateSubclone}{PARAM_DESCRIPTION, Default: TRUE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
