% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excludeCentromere}
\alias{excludeCentromere}
\title{FUNCTION_TITLE}
\usage{
excludeCentromere(x, centromere, flankLength = 0, genomeStyle = "NCBI")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{centromere}{PARAM_DESCRIPTION}

\item{flankLength}{PARAM_DESCRIPTION, Default: 0}

\item{genomeStyle}{PARAM_DESCRIPTION, Default: 'NCBI'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
