% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterByMappabilityScore}
\alias{filterByMappabilityScore}
\title{FUNCTION_TITLE}
\usage{
filterByMappabilityScore(counts, map, mapScoreThres = 0.9)
}
\arguments{
\item{counts}{PARAM_DESCRIPTION}

\item{map}{PARAM_DESCRIPTION}

\item{mapScoreThres}{PARAM_DESCRIPTION, Default: 0.9}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
