% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterByTargetedSequences}
\alias{filterByTargetedSequences}
\title{FUNCTION_TITLE}
\usage{
filterByTargetedSequences(counts, targetedSequences)
}
\arguments{
\item{counts}{PARAM_DESCRIPTION}

\item{targetedSequences}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
