% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{gammapdflog}
\alias{gammapdflog}
\title{FUNCTION_TITLE}
\usage{
gammapdflog(x, a, b)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
