% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{get2and3ComponentMixture}
\alias{get2and3ComponentMixture}
\title{FUNCTION_TITLE}
\usage{
get2and3ComponentMixture(ct, ct.sc.status, n, sp, phi)
}
\arguments{
\item{ct}{PARAM_DESCRIPTION}

\item{ct.sc.status}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}

\item{sp}{PARAM_DESCRIPTION}

\item{phi}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
