% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{get2ComponentMixture}
\alias{get2ComponentMixture}
\title{FUNCTION_TITLE}
\usage{
get2ComponentMixture(ct, n, phi)
}
\arguments{
\item{ct}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}

\item{phi}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
