% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{getDefaultParameters}
\alias{getDefaultParameters}
\title{FUNCTION_TITLE}
\usage{
getDefaultParameters(
  x,
  maxCN = 5,
  ct.sc = NULL,
  ploidy = 2,
  e = 0.9999999,
  e.sameState = 10,
  strength = 1e+07,
  includeHOMD = FALSE
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{maxCN}{PARAM_DESCRIPTION, Default: 5}

\item{ct.sc}{PARAM_DESCRIPTION, Default: NULL}

\item{ploidy}{PARAM_DESCRIPTION, Default: 2}

\item{e}{PARAM_DESCRIPTION, Default: 0.9999999}

\item{e.sameState}{PARAM_DESCRIPTION, Default: 10}

\item{strength}{PARAM_DESCRIPTION, Default: 1e+07}

\item{includeHOMD}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
