% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGender}
\alias{getGender}
\title{FUNCTION_TITLE}
\usage{
getGender(
  rawReads,
  normReads,
  gc,
  map,
  fracReadsInChrYForMale = 0.002,
  chrXMedianForMale = -0.5,
  useChrY = TRUE,
  centromere = NULL,
  flankLength = 1e+05,
  targetedSequences = NULL,
  genomeStyle = "NCBI"
)
}
\arguments{
\item{rawReads}{PARAM_DESCRIPTION}

\item{normReads}{PARAM_DESCRIPTION}

\item{gc}{PARAM_DESCRIPTION}

\item{map}{PARAM_DESCRIPTION}

\item{fracReadsInChrYForMale}{PARAM_DESCRIPTION, Default: 0.002}

\item{chrXMedianForMale}{PARAM_DESCRIPTION, Default: -0.5}

\item{useChrY}{PARAM_DESCRIPTION, Default: TRUE}

\item{centromere}{PARAM_DESCRIPTION, Default: NULL}

\item{flankLength}{PARAM_DESCRIPTION, Default: 1e+05}

\item{targetedSequences}{PARAM_DESCRIPTION, Default: NULL}

\item{genomeStyle}{PARAM_DESCRIPTION, Default: 'NCBI'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
