% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getGenomeWidePositions}
\alias{getGenomeWidePositions}
\title{FUNCTION_TITLE}
\usage{
getGenomeWidePositions(chrs, posns, seqinfo = NULL)
}
\arguments{
\item{chrs}{PARAM_DESCRIPTION}

\item{posns}{PARAM_DESCRIPTION}

\item{seqinfo}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
