% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSeqInfo}
\alias{getSeqInfo}
\title{FUNCTION_TITLE}
\usage{
getSeqInfo(genomeBuild = "hg19", genomeStyle = "NCBI")
}
\arguments{
\item{genomeBuild}{PARAM_DESCRIPTION, Default: 'hg19'}

\item{genomeStyle}{PARAM_DESCRIPTION, Default: 'NCBI'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
