% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{FUNCTION_TITLE}
\usage{
getTransitionMatrix(K, e, strength)
}
\arguments{
\item{K}{PARAM_DESCRIPTION}

\item{e}{PARAM_DESCRIPTION}

\item{strength}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
