% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{HMMsegment}
\alias{HMMsegment}
\title{FUNCTION_TITLE}
\usage{
HMMsegment(
  x,
  validInd = NULL,
  dataType = "copy",
  param = NULL,
  chrTrain = c(1:22),
  maxiter = 50,
  estimateNormal = TRUE,
  estimatePloidy = TRUE,
  estimatePrecision = TRUE,
  estimateSubclone = TRUE,
  estimateTransition = TRUE,
  estimateInitDist = TRUE,
  logTransform = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{validInd}{PARAM_DESCRIPTION, Default: NULL}

\item{dataType}{PARAM_DESCRIPTION, Default: 'copy'}

\item{param}{PARAM_DESCRIPTION, Default: NULL}

\item{chrTrain}{PARAM_DESCRIPTION, Default: c(1:22)}

\item{maxiter}{PARAM_DESCRIPTION, Default: 50}

\item{estimateNormal}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePloidy}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePrecision}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateSubclone}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateTransition}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateInitDist}{PARAM_DESCRIPTION, Default: TRUE}

\item{logTransform}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
