\name{ichorCNA-package}
\alias{ichorCNA-package}
\alias{ichorCNA}
\docType{package}
\title{
\packageTitle{anRpackage}
}
\description{
\packageDescription{anRpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{anRpackage}
\packageIndices{anRpackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{anRpackage}

Maintainer: \packageMaintainer{anRpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
