% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepChr}
\alias{keepChr}
\title{FUNCTION_TITLE}
\usage{
keepChr(tumour_reads, chrs = c(1:22, "X", "Y"))
}
\arguments{
\item{tumour_reads}{PARAM_DESCRIPTION}

\item{chrs}{PARAM_DESCRIPTION, Default: c(1:22, "X", "Y")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
