% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadReadCountsFromWig}
\alias{loadReadCountsFromWig}
\title{FUNCTION_TITLE}
\usage{
loadReadCountsFromWig(
  counts,
  chrs = c(1:22, "X", "Y"),
  gc = NULL,
  map = NULL,
  centromere = NULL,
  flankLength = 1e+05,
  targetedSequences = NULL,
  genomeStyle = "NCBI",
  applyCorrection = TRUE,
  mapScoreThres = 0.9,
  chrNormalize = c(1:22, "X", "Y"),
  fracReadsInChrYForMale = 0.002,
  chrXMedianForMale = -0.5,
  useChrY = TRUE
)
}
\arguments{
\item{counts}{PARAM_DESCRIPTION}

\item{chrs}{PARAM_DESCRIPTION, Default: c(1:22, "X", "Y")}

\item{gc}{PARAM_DESCRIPTION, Default: NULL}

\item{map}{PARAM_DESCRIPTION, Default: NULL}

\item{centromere}{PARAM_DESCRIPTION, Default: NULL}

\item{flankLength}{PARAM_DESCRIPTION, Default: 1e+05}

\item{targetedSequences}{PARAM_DESCRIPTION, Default: NULL}

\item{genomeStyle}{PARAM_DESCRIPTION, Default: 'NCBI'}

\item{applyCorrection}{PARAM_DESCRIPTION, Default: TRUE}

\item{mapScoreThres}{PARAM_DESCRIPTION, Default: 0.9}

\item{chrNormalize}{PARAM_DESCRIPTION, Default: c(1:22, "X", "Y")}

\item{fracReadsInChrYForMale}{PARAM_DESCRIPTION, Default: 0.002}

\item{chrXMedianForMale}{PARAM_DESCRIPTION, Default: -0.5}

\item{useChrY}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
