% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logRbasedCN}
\alias{logRbasedCN}
\title{FUNCTION_TITLE}
\usage{
logRbasedCN(x, purity, ploidyT, cellPrev = NA, cn = 2)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{purity}{PARAM_DESCRIPTION}

\item{ploidyT}{PARAM_DESCRIPTION}

\item{cellPrev}{PARAM_DESCRIPTION, Default: NA}

\item{cn}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
