% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeByPanelOrMatchedNormal}
\alias{normalizeByPanelOrMatchedNormal}
\title{FUNCTION_TITLE}
\usage{
normalizeByPanelOrMatchedNormal(
  tumour_copy,
  chrs = c(1:22, "X", "Y"),
  normal_panel = NULL,
  normal_copy = NULL,
  gender = "female",
  normalizeMaleX = FALSE
)
}
\arguments{
\item{tumour_copy}{PARAM_DESCRIPTION}

\item{chrs}{PARAM_DESCRIPTION, Default: c(1:22, "X", "Y")}

\item{normal_panel}{PARAM_DESCRIPTION, Default: NULL}

\item{normal_copy}{PARAM_DESCRIPTION, Default: NULL}

\item{gender}{PARAM_DESCRIPTION, Default: 'female'}

\item{normalizeMaleX}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
