% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outputHMM}
\alias{outputHMM}
\title{FUNCTION_TITLE}
\usage{
outputHMM(cna, segs, results, patientID = NULL, outDir = ".")
}
\arguments{
\item{cna}{PARAM_DESCRIPTION}

\item{segs}{PARAM_DESCRIPTION}

\item{results}{PARAM_DESCRIPTION}

\item{patientID}{PARAM_DESCRIPTION, Default: NULL}

\item{outDir}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
