% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outputParametersToFile}
\alias{outputParametersToFile}
\title{FUNCTION_TITLE}
\usage{
outputParametersToFile(hmmResults, file)
}
\arguments{
\item{hmmResults}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
