% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotChrLines}
\alias{plotChrLines}
\title{FUNCTION_TITLE}
\usage{
plotChrLines(chrs, chrBkpt, yrange)
}
\arguments{
\item{chrs}{PARAM_DESCRIPTION}

\item{chrBkpt}{PARAM_DESCRIPTION}

\item{yrange}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
