% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCNlogRByChr}
\alias{plotCNlogRByChr}
\title{FUNCTION_TITLE}
\usage{
plotCNlogRByChr(
  dataIn,
  segs,
  param = NULL,
  logR.column = "logR",
  call.column = "event",
  plotSegs = TRUE,
  seqinfo = NULL,
  chr = NULL,
  ploidy = NULL,
  geneAnnot = NULL,
  yrange = c(-4, 6),
  xlim = NULL,
  xaxt = "n",
  cex = 0.5,
  gene.cex = 0.5,
  plot.title = NULL,
  spacing = 4,
  cytoBand = T,
  alphaVal = 1,
  main
)
}
\arguments{
\item{dataIn}{PARAM_DESCRIPTION}

\item{segs}{PARAM_DESCRIPTION}

\item{param}{PARAM_DESCRIPTION, Default: NULL}

\item{logR.column}{PARAM_DESCRIPTION, Default: 'logR'}

\item{call.column}{PARAM_DESCRIPTION, Default: 'event'}

\item{plotSegs}{PARAM_DESCRIPTION, Default: TRUE}

\item{seqinfo}{PARAM_DESCRIPTION, Default: NULL}

\item{chr}{PARAM_DESCRIPTION, Default: NULL}

\item{ploidy}{PARAM_DESCRIPTION, Default: NULL}

\item{geneAnnot}{PARAM_DESCRIPTION, Default: NULL}

\item{yrange}{PARAM_DESCRIPTION, Default: c(-4, 6)}

\item{xlim}{PARAM_DESCRIPTION, Default: NULL}

\item{xaxt}{PARAM_DESCRIPTION, Default: 'n'}

\item{cex}{PARAM_DESCRIPTION, Default: 0.5}

\item{gene.cex}{PARAM_DESCRIPTION, Default: 0.5}

\item{plot.title}{PARAM_DESCRIPTION, Default: NULL}

\item{spacing}{PARAM_DESCRIPTION, Default: 4}

\item{cytoBand}{PARAM_DESCRIPTION, Default: T}

\item{alphaVal}{PARAM_DESCRIPTION, Default: 1}

\item{main}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
