% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCorrectionGenomeWide}
\alias{plotCorrectionGenomeWide}
\title{FUNCTION_TITLE}
\usage{
plotCorrectionGenomeWide(correctOutput, seqinfo = NULL, ...)
}
\arguments{
\item{correctOutput}{PARAM_DESCRIPTION}

\item{seqinfo}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
