% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGWSolution}
\alias{plotGWSolution}
\title{FUNCTION_TITLE}
\usage{
plotGWSolution(
  hmmResults.cor,
  s,
  outPlotFile,
  plotFileType = "pdf",
  logR.column = "logR",
  call.column = "event",
  seqinfo = NULL,
  plotSegs = TRUE,
  plotYLim = c(-2, 2),
  estimateScPrevalence,
  main,
  turnDevOn = TRUE,
  turnDevOff = TRUE
)
}
\arguments{
\item{hmmResults.cor}{PARAM_DESCRIPTION}

\item{s}{PARAM_DESCRIPTION}

\item{outPlotFile}{PARAM_DESCRIPTION}

\item{plotFileType}{PARAM_DESCRIPTION, Default: 'pdf'}

\item{logR.column}{PARAM_DESCRIPTION, Default: 'logR'}

\item{call.column}{PARAM_DESCRIPTION, Default: 'event'}

\item{seqinfo}{PARAM_DESCRIPTION, Default: NULL}

\item{plotSegs}{PARAM_DESCRIPTION, Default: TRUE}

\item{plotYLim}{PARAM_DESCRIPTION, Default: c(-2, 2)}

\item{estimateScPrevalence}{PARAM_DESCRIPTION}

\item{main}{PARAM_DESCRIPTION}

\item{turnDevOn}{PARAM_DESCRIPTION, Default: TRUE}

\item{turnDevOff}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
