% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIchorCNAzoom.R
\name{plotIchorCNAzoom}
\alias{plotIchorCNAzoom}
\title{Runs plotIchorCNAzoom
author: Gavin Ha 
Fred Hutchinson Cancer Research Center
contact: <gha@fredhutch.org>
date:  October 8, 2019
description: Generate plots of copy number from On- and Off-target ichorCNA results}
\usage{
plotIchorCNAzoom(
  id,
  cnFile,
  paramFile,
  geneList = NULL,
  genomeBuild = "hg19",
  genomeStyle = "NCBI",
  zoom = FALSE,
  yaxis = "integer",
  chrs = "c(1:22, 'X')",
  startPos = NULL,
  endPos = NULL,
  plotYlim = "c(-2,2)",
  plotSize = "c(5,3)",
  plotFormat = "png",
  outPlotFile
)
}
\arguments{
\item{id}{Sample ID}

\item{cnFile}{Path to ichorCNA cna.seg output file.}

\item{paramFile}{Path to ichorCNA params.txt output file.}

\item{geneList}{Path to file containing list of genes with chr, start, end coordinates.}

\item{genomeBuild}{Genome build: hg19 or hg38.}

\item{genomeStyle}{NCBI or UCSC chromosome naming convention; use UCSC if desired output is to have \"chr\" string.}

\item{zoom}{Zoom plot; if TRUE, then requires --chrs --start --end to be set.}

\item{yaxis}{Data type to plot for y-axis (\"integer\" copy number or \"logratio\").}

\item{chrs}{Chromosomes to plot; string.}

\item{startPos}{Start coordinate for zoom plots.}

\item{endPos}{End coordinate for zoom plots.}

\item{plotYlim}{Y limits for plotting log ratio.}

\item{plotSize}{width and height in inches.}

\item{plotFormat}{File format of plot. E.g. pdf or png}

\item{outPlotFile}{Path to output figure file.}
}
\description{
Runs plotIchorCNAzoom
author: Gavin Ha 
Fred Hutchinson Cancer Research Center
contact: <gha@fredhutch.org>
date:  October 8, 2019
description: Generate plots of copy number from On- and Off-target ichorCNA results
}
