% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotParam}
\alias{plotParam}
\title{FUNCTION_TITLE}
\usage{
plotParam(mus, lambdas, nu, subclone = NULL, copy.states = 0:6, ...)
}
\arguments{
\item{mus}{PARAM_DESCRIPTION}

\item{lambdas}{PARAM_DESCRIPTION}

\item{nu}{PARAM_DESCRIPTION}

\item{subclone}{PARAM_DESCRIPTION, Default: NULL}

\item{copy.states}{PARAM_DESCRIPTION, Default: 0:6}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
