% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSolutions}
\alias{plotSolutions}
\title{FUNCTION_TITLE}
\usage{
plotSolutions(
  hmmResults.cor,
  tumour_copy,
  logR.column = "logR",
  call.column = "event",
  chrs,
  outDir,
  plotSegs = TRUE,
  numSamples = 1,
  plotFileType = "pdf",
  plotYLim = c(-2, 2),
  seqinfo = NULL,
  estimateScPrevalence = FALSE,
  maxCN
)
}
\arguments{
\item{hmmResults.cor}{PARAM_DESCRIPTION}

\item{tumour_copy}{PARAM_DESCRIPTION}

\item{logR.column}{PARAM_DESCRIPTION, Default: 'logR'}

\item{call.column}{PARAM_DESCRIPTION, Default: 'event'}

\item{chrs}{PARAM_DESCRIPTION}

\item{outDir}{PARAM_DESCRIPTION}

\item{plotSegs}{PARAM_DESCRIPTION, Default: TRUE}

\item{numSamples}{PARAM_DESCRIPTION, Default: 1}

\item{plotFileType}{PARAM_DESCRIPTION, Default: 'pdf'}

\item{plotYLim}{PARAM_DESCRIPTION, Default: c(-2, 2)}

\item{seqinfo}{PARAM_DESCRIPTION, Default: NULL}

\item{estimateScPrevalence}{PARAM_DESCRIPTION, Default: FALSE}

\item{maxCN}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
