% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{priorProbs}
\alias{priorProbs}
\title{FUNCTION_TITLE}
\usage{
priorProbs(
  n,
  sp,
  phi,
  lambda,
  piG,
  A,
  params,
  estimateNormal = TRUE,
  estimatePloidy = TRUE,
  estimatePrecision = TRUE,
  estimateTransition = TRUE,
  estimateInitDist = TRUE,
  estimateSubclone = TRUE
)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{sp}{PARAM_DESCRIPTION}

\item{phi}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{piG}{PARAM_DESCRIPTION}

\item{A}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{estimateNormal}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePloidy}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePrecision}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateTransition}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateInitDist}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateSubclone}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
