% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIchorCNA.R
\name{run_ichorCNA}
\alias{run_ichorCNA}
\title{Runs ichorCNA}
\usage{
run_ichorCNA(
  tumor_wig,
  normal_wig,
  gcWig,
  mapWig,
  repTimeWig,
  normal_panel = NULL,
  sex = NULL,
  exons.bed = NULL,
  id = "test",
  centromere = NULL,
  minMapScore = 0.9,
  flankLength = 1e+05,
  normal = 0.5,
  estimatePloidy = TRUE,
  maxFracCNASubclone = 0.7,
  normal.init = "c(0.5, 0.5)",
  scStates = NULL,
  scPenalty = 0.1,
  normal2IgnoreSC = 1,
  coverage = NULL,
  likModel = "t",
  lambda = NULL,
  lambdaScaleHyperParam = 3,
  kappa = 50,
  ploidy = "2",
  maxCN = 7,
  estimateNormal = TRUE,
  estimateScPrevalence = TRUE,
  maxFracGenomeSubclone = 0.5,
  minSegmentBins = 50,
  altFracThreshold = 0.05,
  chrNormalize = paste0("chr", c(1:22)),
  chrTrain = paste0("chr", c(1:22)),
  chrs = paste0("chr", c(1:22, "X")),
  genomeBuild = "hg19",
  genomeStyle = "UCSC",
  normalizeMaleX = TRUE,
  fracReadsInChrYForMale = 0.001,
  includeHOMD = FALSE,
  txnE = 0.9999999,
  txnStrength = 1e+07,
  multSampleTxnStrength = 1,
  plotFileType = "pdf",
  plotYLim = "c(-2,2)",
  outDir = "./",
  cores = 1
)
}
\arguments{
\item{tumor_wig}{Path to tumor WIG file. Required.}

\item{normal_wig}{Path to normal WIG file.}

\item{gcWig}{Path to GC-content WIG file; Required.}

\item{mapWig}{Path to mappability score WIG file.}

\item{repTimeWig}{Path to replication timing WIG file.}

\item{normal_panel}{Median corrected depth from panel of normals.}

\item{sex}{User specified gender: male or female}

\item{exons.bed}{Path to bed file containing exon regions.}

\item{id}{Patient ID.}

\item{centromere}{File containing Centromere locations; if not provided then will use hg19 version from ichorCNA package.}

\item{minMapScore}{Include bins with a minimum mappability score of this value.}

\item{flankLength}{Length of region flanking centromere to remove.}

\item{normal}{Initial normal contamination; can be more than one value if additional normal initializations are desired.}

\item{estimatePloidy}{Estimate tumour ploidy.}

\item{maxFracCNASubclone}{Exclude solutions with fraction of subclonal events greater than this value.}

\item{normal.init}{Specific initialization of normal contamination for multiple samples.}

\item{scStates}{Subclonal states to consider.}

\item{scPenalty}{Penalty for subclonal state transitions, 0.1 penalizes subclonal states by ~10 percent.}

\item{normal2IgnoreSC}{Ignore subclonal analysis when normal proportion is greater than this value.}

\item{coverage}{PICARD sequencing coverage.}

\item{likModel}{Likelihood model to use: "t" or "gaussian". Use "gaussian" for faster runtimes.}

\item{lambda}{Initial Student's t precision; must contain 4 values (e.g. c(1500,1500,1500,1500)); if not provided then will automatically use based on variance of data.}

\item{lambdaScaleHyperParam}{Hyperparameter (scale) for Gamma prior on Student's-t precision.}

\item{kappa}{Initial state distribution")}

\item{ploidy}{Initial tumour ploidy; can be more than one value if additional ploidy initializations are desired.}

\item{maxCN}{Total clonal CN states.}

\item{estimateNormal}{Estimate normal.}

\item{estimateScPrevalence}{Estimate subclonal prevalence.}

\item{maxFracGenomeSubclone}{Exclude solutions with subclonal genome fraction greater than this value.}

\item{minSegmentBins}{Minimum number of bins for largest segment threshold required to estimate tumor fraction if below this threshold, then will be assigned zero tumor fraction.}

\item{altFracThreshold}{Minimum proportion of bins altered required to estimate tumor fraction; if below this threshold, then will be assigned zero tumor fraction.}

\item{chrNormalize}{Specify chromosomes to normalize GC/mappability biases.}

\item{chrTrain}{Specify chromosomes to estimate params.}

\item{chrs}{Specify chromosomes to analyze.}

\item{genomeBuild}{Geome build.}

\item{genomeStyle}{NCBI or UCSC chromosome naming convention; use UCSC if desired output is to have "chr" string.}

\item{normalizeMaleX}{If male, then normalize chrX by median.}

\item{fracReadsInChrYForMale}{Threshold for fraction of reads in chrY to assign as male.}

\item{includeHOMD}{If FALSE, then exclude HOMD state; Useful when using large bins (e.g. 1Mb).}

\item{txnE}{Self-transition probability; Increase to decrease number of segments.}

\item{txnStrength}{Transition pseudo-counts; Exponent should be the same as the number of decimal places of `txnE`.}

\item{multSampleTxnStrength}{Strength of same state transition between multiple samples.}

\item{plotFileType}{File format for output plots.}

\item{plotYLim}{Ylim to use for chromosome plots.}

\item{outDir}{Output Directory.}

\item{cores}{Number of cores to use for EM.}
}
\description{
Runs ichorCNA
}
