% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{runEM}
\alias{runEM}
\title{FUNCTION_TITLE}
\usage{
runEM(
  copy,
  chr,
  chrTrain,
  param,
  maxiter,
  verbose = TRUE,
  estimateNormal = TRUE,
  estimatePloidy = TRUE,
  estimatePrecision = TRUE,
  estimateTransition = TRUE,
  estimateInitDist = TRUE,
  estimateSubclone = TRUE,
  likChangeConvergence = 0.001
)
}
\arguments{
\item{copy}{PARAM_DESCRIPTION}

\item{chr}{PARAM_DESCRIPTION}

\item{chrTrain}{PARAM_DESCRIPTION}

\item{param}{PARAM_DESCRIPTION}

\item{maxiter}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateNormal}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePloidy}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimatePrecision}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateTransition}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateInitDist}{PARAM_DESCRIPTION, Default: TRUE}

\item{estimateSubclone}{PARAM_DESCRIPTION, Default: TRUE}

\item{likChangeConvergence}{PARAM_DESCRIPTION, Default: 0.001}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
