% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{runIchorCNA}
\alias{runIchorCNA}
\title{Run IchorCNA}
\usage{
runIchorCNA(
  WIG = NULL,
  NORMWIG = NULL,
  gcWig = NULL,
  mapWig = NULL,
  normalPanel = NULL,
  exons.bed = NULL,
  patientID = "test",
  centromere = NULL,
  minMapScore = 0.9,
  rmCentromereFlankLength = 1e+05,
  normal = 0.5,
  scStates = NULL,
  coverage = NULL,
  lambda = NULL,
  lambdaScaleHyperParam = 3,
  ploidy = 2,
  maxCN = 7,
  estimateNormal = TRUE,
  estimateScPrevalence = TRUE,
  estimatePloidy = TRUE,
  maxFracCNASubclone = 0.7,
  maxFracGenomeSubclone = 0.5,
  minSegmentBins = 50,
  altFracThreshold = 0.05,
  chrNormalize = c(1:22),
  chrTrain = c(1:22),
  chrs = c(1:22, "X"),
  genomeBuild = "hg19",
  genomeStyle = "NCBI",
  normalizeMaleX = TRUE,
  minTumFracToCorrect = 0.1,
  fracReadsInChrYForMale = 0.001,
  includeHOMD = FALSE,
  txnE = 0.9999999,
  txnStrength = 1e+07,
  plotFileType = "pdf",
  plotYLim = c(-2, 2),
  outDir = "./",
  libdir = NULL
)
}
\arguments{
\item{libdir}{}
}
\value{

}
\description{
Run IchorCNA
}
