% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{runViterbi}
\alias{runViterbi}
\title{FUNCTION_TITLE}
\usage{
runViterbi(convergedParams, chr)
}
\arguments{
\item{convergedParams}{PARAM_DESCRIPTION}

\item{chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
