% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentData}
\alias{segmentData}
\title{FUNCTION_TITLE}
\usage{
segmentData(dataGR, validInd, states, convergedParams)
}
\arguments{
\item{dataGR}{PARAM_DESCRIPTION}

\item{validInd}{PARAM_DESCRIPTION}

\item{states}{PARAM_DESCRIPTION}

\item{convergedParams}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
