% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setGenomeStyle}
\alias{setGenomeStyle}
\title{FUNCTION_TITLE}
\usage{
setGenomeStyle(x, genomeStyle = "NCBI", species = "Homo_sapiens")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{genomeStyle}{PARAM_DESCRIPTION, Default: 'NCBI'}

\item{species}{PARAM_DESCRIPTION, Default: 'Homo_sapiens'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
