% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{stLikelihood}
\alias{stLikelihood}
\title{FUNCTION_TITLE}
\usage{
stLikelihood(n, sp, phi, lambda, params, D, rho)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{sp}{PARAM_DESCRIPTION}

\item{phi}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{D}{PARAM_DESCRIPTION}

\item{rho}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
