% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{tdistPDF}
\alias{tdistPDF}
\title{FUNCTION_TITLE}
\usage{
tdistPDF(x, mu, lambda, nu)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{mu}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{nu}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
