% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wigToGRanges}
\alias{wigToGRanges}
\title{FUNCTION_TITLE}
\usage{
wigToGRanges(wigfile, verbose = TRUE)
}
\arguments{
\item{wigfile}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
