\name{ichorCNA-package}
\alias{ichorCNA-package}
\alias{ichorCNA}
\docType{package}
\title{
ichorCNA
}
\description{
ichorCNA is a tool for estimating the fraction of tumor in ultra-low-pass whole genome sequencing (ULP-WGS) of cell-free DNA.
}
\details{

\tabular{ll}{
Package: \tab ichorCNA\cr
Type: \tab Package\cr
Version: \tab 0.3.1\cr
Date: \tab 2019-07-12\cr
License: \tab GPL-3\cr
}
}

\references{
Adalsteinsson, VA*, Ha G*, Freeman SS*, Choudhury AD, Stover DG, Parsons HA, Gydush G, Reed SC, Rotem D, Rhoades J, Loginov D, Livitz D, Rosebrock D, Leshchiner I, Kim J, Stewart C, Rosenberg M, Francis JM, Zhang CZ, Cohen O, Oh C, Ding H, Polak Paz, Lloyd M, Mahmud S, Helvie S, Merrill MS, Santiago RA, O’Connor EP, Jeong SH, Leeson R, Barry RM, Kramkowski JF, Zhang Z, Polacek L, Lohr J, Schleicher M, Lipscomb E, Saltzman A, Oliver NM, Marini L, Waks AG, Harshman LC, Tolaney SM,  Van Allen EM, Winer EP, Lin NU, Nakabayashi M, Taplin ME, Johannessen CM, Garraway LA, Golub RE, Boehm JS, Wagle N, Getz G*, Love JC*, Meyerson M*. Scalable whole-exome sequencing of cell-free DNA reveals high concordance with metastatic tumors. Nature Communications 2017 8:1324. doi: 10.1038/s41467-017-00965-y
}

\author{
Gavin Ha <gha@fredhutch.org>
Justin Rhoades <rhoades@broadinstitute.org>
Samuel Freeman <sfreeman@broadinstitute.org>
}

\examples{

}

